requirements.txt
apache-airflow==2.3.4
apache-airflow-providers-google==8.11.0
google-cloud-bigquery-reservation==1.7.0